Revert "Quality"
