+++ src/models.py
@@ -5,11 +5,12 @@ import torch
 class UNetDown(nn.Module):
     def __init__(self, in_size, out_size, normalize=True, dropout=0.0):
         super(UNetDown, self).__init__()
-        layers = [nn.Conv2d(in_size, out_size, 4, 2, 1, bias=True),
-                  nn.BatchNorm2d(out_size, momentum=0.8),
-                  nn.LeakyReLU(0.2)]
-
-        # if dropout:
+        layers = [nn.Conv2d(in_size, out_size, 4, 2, 1, bias=True)]
+        if normalize:
+            #layers.append(nn.InstanceNorm2d(out_size))
+            layers.append(nn.BatchNorm2d(out_size, momentum=0.8))
+        layers.append(nn.LeakyReLU(0.2))
+        #if dropout:
         #    layers.append(nn.Dropout(dropout))
         self.model = nn.Sequential(*layers)
 
@@ -22,10 +23,12 @@ class UNetUp(nn.Module):
         super(UNetUp, self).__init__()
         layers = [
             nn.ConvTranspose2d(in_size, out_size, 4, 2, 1, bias=True),
-            nn.Conv2d(out_size, out_size, 3, 1, 1),  # ???
+            nn.Conv2d(out_size, out_size, 3, 1, 1),
+            #nn.InstanceNorm2d(out_size),
             nn.BatchNorm2d(out_size, momentum=0.8),
-            nn.LeakyReLU(inplace=True)]
-        # if dropout:
+            nn.ReLU(inplace=True),
+        ]
+        #if dropout:
         #    layers.append(nn.Dropout(dropout))
 
         self.model = nn.Sequential(*layers)
@@ -33,6 +36,7 @@ class UNetUp(nn.Module):
     def forward(self, x, skip_input):
         x = self.model(x)
         x = torch.cat((x, skip_input), 1)
+
         return x
 
 
@@ -43,49 +47,43 @@ class GeneratorUNet(nn.Module):
         self.down1 = UNetDown(in_channels, 64, normalize=False)
         self.down2 = UNetDown(64, 128)
         self.down3 = UNetDown(128, 256)
-        self.down4 = UNetDown(256, 512)  # , dropout=0.5)
-        self.down5 = UNetDown(512, 512)  # , dropout=0.5)
-        self.down6 = UNetDown(512, 512)  # , dropout=0.5)
-        self.down7 = UNetDown(512, 512)  # , dropout=0.5)
-        self.down8 = UNetDown(512, 512)  # , dropout=0.5)
+        self.down4 = UNetDown(256, 512) #, dropout=0.5)
+        self.down5 = UNetDown(512, 512)#, dropout=0.5)
+        self.down6 = UNetDown(512, 512)#, dropout=0.5)
+        self.down7 = UNetDown(512, 512)#, dropout=0.5)
+        #self.down8 = UNetDown(512, 512)#, dropout=0.5)
 
-        self.up1 = UNetUp(512 + 3, 512)  # , dropout=0.5)
-        self.up2 = UNetUp(1024, 512)  # , dropout=0.5)
-        self.up3 = UNetUp(1024, 512)  # , dropout=0.5)
-        self.up4 = UNetUp(1024, 512)  # , dropout=0.5)
+        #self.up1 = UNetUp(512, 512)#, dropout=0.5)
+        self.up2 = UNetUp(512, 512)#, dropout=0.5)
+        self.up3 = UNetUp(1024, 512)#, dropout=0.5)
+        self.up4 = UNetUp(1024, 512)#, dropout=0.5)
         self.up5 = UNetUp(1024, 256)
         self.up6 = UNetUp(512, 128)
         self.up7 = UNetUp(256, 64)
 
+
         self.final = nn.Sequential(
             nn.ConvTranspose2d(128, out_channels, 4, 2, 1),
-            # nn.Upsample(scale_factor=2),
-            # nn.ZeroPad2d((1, 0, 1, 0)),
+            #nn.Upsample(scale_factor=2),
+            #nn.ZeroPad2d((1, 0, 1, 0)),
             nn.Conv2d(out_channels, out_channels, 3, 1, 1),
-            # nn.Tanh(),
+            #nn.Tanh(),
             nn.Sigmoid(),
         )
 
-    def forward(self, x, quality):
+    def forward(self, x):
         # U-Net generator with skip connections from encoder to decoder
-
         d1 = self.down1(x)
         d2 = self.down2(d1)
         d3 = self.down3(d2)
         d4 = self.down4(d3)
         d5 = self.down5(d4)
-        d6 = self.down6(d5)  # [16, 512, 4, 4]
-        d7 = self.down7(d6)  # [16, 512, 2, 2]
-        d8 = self.down8(d7)  # [16, 512, 1, 1]
-
-        quality = torch.transpose(quality, 1, 2).unsqueeze(3)
-        d8 = torch.cat((d8, quality), 1)  # [16, 515, 1, 1]
-
-        u1 = self.up1(d8, d7)
-        u2 = self.up2(u1, d6)
-
+        d6 = self.down6(d5)
+        d7 = self.down7(d6)
+        #d8 = self.down8(d7)
+        #u1 = self.up1(d8, d7)
+        u2 = self.up2(d7, d6)
         u3 = self.up3(u2, d5)
-
         u4 = self.up4(u3, d4)
         u5 = self.up5(u4, d3)
         u6 = self.up6(u5, d2)
@@ -105,10 +103,11 @@ class Discriminator(nn.Module):
 
         def discriminator_block(in_filters, out_filters, normalization=True):
             """Returns downsampling layers of each discriminator block"""
-            layers = [nn.Conv2d(in_filters, out_filters, 4, stride=2, padding=1),
-                      nn.BatchNorm2d(out_filters),
-                      nn.LeakyReLU(0.2, inplace=True)]
-
+            layers = [nn.Conv2d(in_filters, out_filters, 4, stride=2, padding=1)]
+            if normalization:
+                #layers.append(nn.InstanceNorm2d(out_filters))
+                layers.append(nn.BatchNorm2d(out_filters))
+            layers.append(nn.LeakyReLU(0.2, inplace=True))
             return layers
 
         self.model = nn.Sequential(
@@ -120,7 +119,7 @@ class Discriminator(nn.Module):
             nn.Conv2d(512, 1, 4, padding=1, bias=False)
         )
 
-    def forward(self, condition, target):  # [16, 1, 256, 256], [16, 1, 256, 256]
+    def forward(self, condition, target):
         # Concatenate image and condition image by channels to produce input
-        img_input = torch.cat((condition, target), 1)  # [16, 2, 256, 256]
+        img_input = torch.cat((condition, target), 1)
         return self.model(img_input)
